backend/requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.0.post1
pydantic>=2.5.3
pydantic-settings>=2.1.0
httpx>=0.26.0
python-dotenv>=1.0.1
sse-starlette>=2.0.0
tenacity>=8.2.3
loguru>=0.7.2
slowapi<=0.1.9
faiss-cpu>=1.7.4
sentence-transformers>=2.3.1
jinja2>=3.1.3
